Correccion nombramiento de clases Controller. Ajustes negocio AmbienteService.
