+++ src/main/java/com/asignacion/espacios/controller/ModuloController.java
@@ -1,5 +1,6 @@
 package com.asignacion.espacios.controller;
 
+
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
@@ -21,60 +22,66 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.asignacion.espacios.clases.Mensaje;
+import com.asignacion.espacios.model.dto.AmbienteDTO;
+import com.asignacion.espacios.model.entity.MaeGrupoLista;
 import com.asignacion.espacios.model.entity.Modulo;
-import com.asignacion.espacios.model.entity.Puesto;
+import com.asignacion.espacios.service.IAmbienteService;
+import com.asignacion.espacios.service.IMaeGrupoListaService;
 import com.asignacion.espacios.service.IModuloService;
-import com.asignacion.espacios.service.IPuestoService;
+
 
 @Controller
-@RequestMapping("/puesto")
-public class puestoController {
-
-	@Autowired
-	IPuestoService servicePuesto;
+@RequestMapping("/modulo")
+public class ModuloController {
 	
-
 	@Autowired
 	IModuloService serviceModulo;
+	
+	@Autowired
+	IMaeGrupoListaService serviceIMaeGrupoLista;
+	
+	@Autowired
+	IAmbienteService serviceAmbiente;
+
 
 	
-	@GetMapping("/listarPuesto")
-	public String listarPuesto(Model model) {
-		List<Puesto> listaPuestos = servicePuesto.listarTodos();
-		model.addAttribute("listaPuestos", listaPuestos);
-		return "puesto/listarPuesto";
-	}
-	
-	@GetMapping("/crearPuesto")
-	public String crearPuesto(Model model, Puesto puesto) {
+	@GetMapping("/listarModulos")
+	public String listarModulos(Model model) {
 		List<Modulo> listaModulos = serviceModulo.listarTodos();
-		model.addAttribute("listaModulos",listaModulos);
-		model.addAttribute("Puesto", puesto);
-		return "puesto/crearPuesto";
+		model.addAttribute("listaModulos", listaModulos);
+		System.out.println("moduloController.listarModulos()");
+		return "modulo/listarModulos";
+	}
+	
+	@GetMapping("/crearModulo")
+	public String crearModulo(Model model, Modulo modulo) {
+		// List<Ambiente>listaAmbiente = serviceAmbiente.listarTodos();
+		// model.addAttribute("listaAmbiente", listaAmbiente);
+		model.addAttribute("Modulo", modulo);
+		return "modulo/crearModulo";
+	}
+	
+	@GetMapping("/editarModulo/{idModulo}")
+	public String editarModulo(Model model, Modulo modulo, @PathVariable("idModulo")int idModulo) {
+		modulo = serviceModulo.buscarModuloId(idModulo);
+		model.addAttribute("Modulo", modulo);
+		// List<Ambiente>listaAmbiente = serviceAmbiente.listarTodos();
+		// model.addAttribute("listaAmbientes", listaAmbiente);
+		return "modulo/crearModulo";
 	}
 	
 	
-	@GetMapping("/editarPuesto/{idPuesto}")
-	public String editarPuesto(Model model, Puesto puesto, @PathVariable("idPuesto")int idPuesto) {
-		puesto = servicePuesto.buscarPuestoId(idPuesto);
-		model.addAttribute("Puesto", puesto);
-		List<Modulo> listaModulos = serviceModulo.listarTodos();
-	        model.addAttribute("listaModulos", listaModulos);
-		return "puesto/crearPuesto";
-	}
-	
-	
-	@PostMapping("/guardarPuesto")
-	public String guardarPuesto (Model model, Puesto puesto, RedirectAttributes attributes) {
+	@PostMapping("/guardarModulo")
+	public String guardarModulo (Model model, Modulo modulo, RedirectAttributes attributes) {
 		Mensaje mensaje = new Mensaje();
-		mensaje = servicePuesto.guardarValidando(puesto);
+		mensaje = serviceModulo.guardarValidando(modulo);
 		if(mensaje.getCodigoMensaje()==0) {
-			model.addAttribute("Puesto", puesto);
+			model.addAttribute("Modulo", modulo);
 			model.addAttribute(mensaje.getTipoAlerta(), mensaje.getDescripcionMensaje());
-			return "puesto/crearPuesto";
+			return "modulo/crearModulo";
 		}
 		attributes.addFlashAttribute(mensaje.getTipoAlerta(), mensaje.getDescripcionMensaje());
-		return "redirect:/puesto/listarPuesto";
+		return "redirect:/modulo/listarModulos";
 	}
 
 	
@@ -101,7 +108,14 @@ public class puestoController {
 		//Listar maestras
 		@ModelAttribute
 		public void setGenericos(Model model, Authentication auth) {
-			
+			  List<MaeGrupoLista> listaOrientacion = serviceIMaeGrupoLista.listaOpcionesHabilitadasPorIdGrupoOrdenadosOrden(3);
+			    model.addAttribute("listaOrientacion", listaOrientacion);
+			    List<MaeGrupoLista> listaDistribucion = serviceIMaeGrupoLista.listaOpcionesHabilitadasPorIdGrupoOrdenadosOrden(8);
+			    model.addAttribute("listaDistribucion", listaDistribucion);
+			    List<AmbienteDTO> listaAmbientes = serviceAmbiente.listarTodos();
+			    model.addAttribute("listaAmbientes", listaAmbientes);
+	
 		}
-
+	
+	
 }
