+++ src/main/java/com/asignacion/espacios/model/dto/AmbienteDTO.java
@@ -6,9 +6,25 @@ import java.util.Date;
 public class AmbienteDTO implements Serializable {
 
 	private static final long serialVersionUID = -1726079802252266181L;
+	private Integer idAmbiente;
 	private String descripcion;
 	private Date fechaCreacion;
 	private boolean indHabilitado;
+
+	public AmbienteDTO(String descripcion, Date fechaCreacion, boolean indHabilitado) {
+		super();
+		this.descripcion = descripcion;
+		this.fechaCreacion = fechaCreacion;
+		this.indHabilitado = indHabilitado;
+	}
+
+	public Integer getIdAmbiente() {
+		return idAmbiente;
+	}
+
+	public void setIdAmbiente(Integer idAmbiente) {
+		this.idAmbiente = idAmbiente;
+	}
 
 	public String getDescripcion() {
 		return descripcion;
