+++ src/main/java/com/asignacion/espacios/service/IAmbienteService.java
@@ -3,20 +3,22 @@ package com.asignacion.espacios.service;
 import java.util.List;
 
 import com.asignacion.espacios.clases.Mensaje;
+import com.asignacion.espacios.model.dto.AmbienteDTO;
 import com.asignacion.espacios.model.entity.Ambiente;
 
-
 public interface IAmbienteService {
-	
+
 	// NATIVO=======================================================================================================================
 	Ambiente buscarAmbienteId(Integer idAmbiente);
-	List<Ambiente> listarTodos();
-	void guardarAmbiente(Ambiente ambiente);
+
+	List<AmbienteDTO> listarTodos();
+
+	void guardarAmbiente(AmbienteDTO ambiente);
 
 	// REPOSITORY===================================================================================================================
-	
-	// METODOS Y LOGICA=============================================================================================================
-	Mensaje guardarValidando(Ambiente ambiente);
 
+	// METODOS Y
+	// LOGICA=============================================================================================================
+	Mensaje guardarValidando(AmbienteDTO ambiente);
 
 }
