+++ src/main/java/com/asignacion/espacios/service/impl/ClinicaServiceImp.java
@@ -1,4 +1,4 @@
-package com.asignacion.espacios.serviceImp;
+package com.asignacion.espacios.service.impl;
 
 import java.util.List;
 import java.util.Optional;
