+++ src/main/java/com/asignacion/espacios/service/impl/PersonaServiceImp.java
@@ -1,4 +1,4 @@
-package com.asignacion.espacios.serviceImp;
+package com.asignacion.espacios.service.impl;
 
 import java.util.ArrayList;
 import java.util.List;
